Moved yes no prompt to DungeonMaster. Also added ability to add characters that were tried to be checked, but didn't exist
